ref: introduce processing run boundary
